+++ babel.config.js
@@ -14,7 +14,8 @@ module.exports = {
           '@services': './src/services',
           '@store': './src/store',
           '@utils': './src/utils',
-          '@ledgerhq/devices/hid-framing': '@ledgerhq/devices/lib-es/hid-framing',
+          '@ledgerhq/devices/hid-framing':
+            '@ledgerhq/devices/lib-es/hid-framing',
         },
       },
     ],
