+++ src/constants/index.ts
@@ -6,3 +6,4 @@ export const MNEMONIC_LENGTH = 24;
 export const API_URL = 'https://testnet-api.multiversx.com';
 export const EEXPLORER_LINK =
   'https://testnet-explorer.multiversx.com/transactions';
+export const DAPP_URL = 'http://192.168.100.86:3000/';
