+++ src/features/DApp/Dapp.tsx
@@ -1,33 +1,14 @@
-import React, { useEffect } from 'react';
-import WebView, { WebViewMessageEvent } from 'react-native-webview';
-import { useGetAccount, useGetAccountInfo } from '@multiversx/sdk-dapp/hooks';
+import { useWalletStore } from '@store/walletStore/walletStore';
+import React from 'react';
+import DappWebviewWrapper from './DappWebviewWrapper/DappWebviewWrapper';
+import AccountNotConnected from './AccountNotConnected/AccountNotConnected';
+import { observer } from 'mobx-react-lite';
 
 const Dapp = () => {
-  const webviewRef = React.useRef<WebView>(null);
+  const { walletStore } = useWalletStore();
+  const { secretKey } = walletStore;
 
-  // const getAcctount = useGetAccount();
-  // const getAcctountInfo = useGetAccountInfo();
-
-  // console.log({ getAcctount, getAcctountInfo });
-
-  const handleOnMessage = (event: WebViewMessageEvent) => {
-    console.log('======= START');
-    console.log({ test: event.nativeEvent.data });
-    setTimeout(() => {
-      webviewRef.current?.postMessage('Hello from react native');
-    }, 2000);
-
-    console.log('======= END');
-  };
-
-  return (
-    <WebView
-      ref={webviewRef}
-      source={{ uri: 'http://192.168.100.86:3000/' }}
-      style={{ flex: 1 }}
-      onMessage={handleOnMessage}
-    />
-  );
+  return secretKey ? <DappWebviewWrapper /> : <AccountNotConnected />;
 };
 
-export default Dapp;
+export default observer(Dapp);
